requirements.txt
python-telegram-bot>=20.0
asyncpg>=0.28.0